+++ src/auto_subtitle/__init__.py
@@ -1,4 +1,4 @@
 from .cli import main
 
 __all__ = ["main"]
-__version__ = "1.0.0"
+__version__ = "1.0.1"